+++ test/fuzz/InvariantTesting.t.sol
@@ -14,9 +14,9 @@ contract InvarianTesting is StdInvariant, Test, IZeroDay {
     string public tokenURIHash = "bafybeibc5sgo2plmjkq2tzmhrn54bk3crhnc23zd2msg4ea7a4pxrkgfna";
     ZeroDay public nft;
     uint256 public constant init_pre_sale_price_example = 1 ether;
-    uint256 public constant start_pre_sale_date_example = 1743493497; // Tuesday, April 1, 2025 7:44:57 AM
-    uint256 public constant start_reveal_date_example = 1746085497; // Thursday, May 1, 2025 7:44:57 AM
-    uint256 public constant start_public_sale_date_example = 1748763897; //Sunday, June 1, 2025 7:44:57 AM
+    uint32 public constant start_pre_sale_date_example = 1743493497; // Tuesday, April 1, 2025 7:44:57 AM
+    uint32 public constant start_reveal_date_example = 1746085497; // Thursday, May 1, 2025 7:44:57 AM
+    uint32 public constant start_public_sale_date_example = 1748763897; //Sunday, June 1, 2025 7:44:57 AM
 
     uint256 public constant PUBLIC_SALE_MINT_PRICE = 1 ether;
     uint96 public constant ROYALTY_BASIS_POINT_VALUE = 500; // 5% of token Royalty.
